Move per-record processing into by_record module out of by_page
